pCRM_appExchangeSplashLanding
PropertyCRM- App Exchange: "Landing Page / Splash Page" 
Contains all pertinent files for this initial shell of page & multiple app container options.
"App containers / tiles" are based on the given Balsamiq Mockup for this page.
There are 5 variations of boxes, 2 of which have been approved by Ree Chen.
Archive.zip is the archive file. 
There is more to be done on this, hoping to circle back for creative adjustments. First, CRM styling & 'shell' will be added to this per CRM's PSDs.